исправлена ошибка в тесте

+++ include/tbitfield.h
@@ -12,7 +12,7 @@
 
 using namespace std;
 
-typedef unsigned int TELEM;
+typedef unsigned long long int TELEM;
 
 class TBitField
 {
